docs : controller 권한에 따라 분리 및 interceptor 반영
